refactor events utils in e2e
