+++ test/e2e/common/BUILD
@@ -17,7 +17,6 @@ go_library(
         "downward_api.go",
         "downwardapi_volume.go",
         "empty_dir.go",
-        "events.go",
         "expansion.go",
         "host_path.go",
         "init_container.go",
@@ -73,6 +72,7 @@ go_library(
         "//staging/src/k8s.io/client-go/tools/cache:go_default_library",
         "//staging/src/k8s.io/client-go/tools/watch:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/framework/events:go_default_library",
         "//test/e2e/framework/kubelet:go_default_library",
         "//test/e2e/framework/network:go_default_library",
         "//test/e2e/framework/node:go_default_library",
