+++ test/e2e/framework/BUILD
@@ -112,6 +112,7 @@ filegroup(
         "//test/e2e/framework/deployment:all-srcs",
         "//test/e2e/framework/deviceplugin:all-srcs",
         "//test/e2e/framework/endpoints:all-srcs",
+        "//test/e2e/framework/events:all-srcs",
         "//test/e2e/framework/ginkgowrapper:all-srcs",
         "//test/e2e/framework/gpu:all-srcs",
         "//test/e2e/framework/ingress:all-srcs",
